[2] refactor: isolate event logic in files
