--- src/lib/index.js
@@ -1,1 +1,0 @@
-export { spy } from './spy.js';
