+++ src/lib/index.ts
@@ -1,0 +1,2 @@
+export { spy } from './action/spy.js';
+export type { SignalEvent, SignalEventDetail } from './events/signal.js';
